move voice commands into separate python file
